requirements.txt
requests
urllib3>=1.26
Flask>=2,<3
python-dotenv>=1.0,<2
python-telegram-bot>=20,<21
